requirements.txt
streamlit
fastapi
uvicorn
jinja2
google-generativeai
Pillow
python-multipart